requirements.txt
openai>=1.30.0
jinja2>=3.1.0
weasyprint>=62.0
python-dotenv>=1.0.0
requests>=2.31.0
python-telegram-bot>=21.0
